Reworked comments generation
Fully reworked comments generation and new comment genration. There are still some problems, but I'm looking for solutions.
